+++ py/desispec/scripts/stdstars.py
@@ -18,6 +18,7 @@ from desispec.interpolation import resample_flux
 from desiutil.log import get_logger
 from desispec.parallel import default_nproc
 from desispec.io.filters import load_filter
+from desispec.dustextinction import ext_odonnell
 
 def parse(options=None):
     parser = argparse.ArgumentParser(description="Extract spectra from pre-processed raw data.")
@@ -30,11 +31,12 @@ def parse(options=None):
     parser.add_argument('--starmodels', type = str, help = 'path of spectro-photometric stellar spectra fits')
     parser.add_argument('-o','--outfile', type = str, help = 'output file for normalized stdstar model flux')
     parser.add_argument('--ncpu', type = int, default = default_nproc, required = False, help = 'use ncpu for multiprocessing')
-    parser.add_argument('--delta-color', type = float, default = 0.1, required = False, help = 'max delta-color for the selection of standard stars (on top of meas. errors)')
+    parser.add_argument('--delta-color', type = float, default = 0.2, required = False, help = 'max delta-color for the selection of standard stars (on top of meas. errors)')
     parser.add_argument('--color', type = str, default = "G-R", required = False, help = 'color for selection of standard stars')
-    parser.add_argument('--z-max', type = float, default = 0.005, required = False, help = 'max peculiar velocity (blue/red)shift range')
-    parser.add_argument('--z-res', type = float, default = 0.00005, required = False, help = 'dz grid resolution')
-
+    parser.add_argument('--z-max', type = float, default = 0.008, required = False, help = 'max peculiar velocity (blue/red)shift range')
+    parser.add_argument('--z-res', type = float, default = 0.00002, required = False, help = 'dz grid resolution')
+    parser.add_argument('--template-error', type = float, default = 0.1, required = False, help = 'fractional template error used in chi2 computation (about 10% for BOSS b1)')
+    
     args = None
     if options is None:
         args = parser.parse_args()
@@ -77,6 +79,11 @@ def get_color(mags,filters,color_name) :
         log.warning("cannot compute '%s' color from %s"%(color_name,filters))
         return 0.
     return mags[index1]-mags[index2]
+
+def dust_transmission(wave,ebv) :
+    Rv = 3.1
+    extinction = ext_odonnell(wave,Rv=Rv)
+    return 10**(-Rv*extinction*ebv/2.5)
 
 def main(args) :
     """ finds the best models of all standard stars in the frame
@@ -122,26 +129,19 @@ def main(args) :
             log.error("incompatible fibermap")
             raise ValueError("incompatible fibermap")
 
-        if camera in frames:
-            log.error("cannot handle for now several frame of same camera (%s)"%camera)
-            raise ValueError("cannot handle for now several frame of same camera (%s)"%camera)
-
-        frames[camera]=frame
-
+        if not camera in frames :
+            frames[camera]=[]
+        frames[camera].append(frame)
+ 
     for filename in args.skymodels :
         log.info("reading %s"%filename)
         sky=io.read_sky(filename)
         header=fits.getheader(filename, 0)
         camera=safe_read_key(header,"CAMERA").strip().lower()
-
-        # NEED TO ADD MORE CHECKS
-        if camera in skies:
-            log.error("cannot handle several skymodels of same camera (%s)"%camera)
-            raise ValueError("cannot handle several skymodels of same camera (%s)"%camera)
-
-
-        skies[camera]=sky
-
+        if not camera in skies :
+            skies[camera]=[]
+        skies[camera].append(sky)
+        
     for filename in args.fiberflats :
         log.info("reading %s"%filename)
         header=fits.getheader(filename, 0)
@@ -150,10 +150,12 @@ def main(args) :
 
         # NEED TO ADD MORE CHECKS
         if camera in flats:
-            log.error("cannot handle several flats of same camera (%s)"%camera)
-            raise ValueError("cannot handle several flats of same camera (%s)"%camera)
-        flats[camera]=flat
 
+            log.warning("cannot handle several flats of same camera (%s), will use only the first one"%camera)
+            #raise ValueError("cannot handle several flats of same camera (%s)"%camera)
+        else :
+            flats[camera]=flat
+    
 
     if starindices.size == 0 :
         log.error("no STD star found in fibermap")
@@ -166,9 +168,15 @@ def main(args) :
     imaging_mags=fibermap["MAG"][starindices]
 
     log.warning("NO MAG ERRORS IN FIBERMAP, I AM IGNORING MEASUREMENT ERRORS !!")
-    log.warning("NO EXTINCTION VALUES IN FIBERMAP, I AM IGNORING THIS FOR NOW !!")
 
-
+    ebv=np.zeros(starindices.size)
+    if "SFD_EBV" in fibermap.columns.names  :
+        log.info("Using 'SFD_EBV' from fibermap")
+        ebv=fibermap["SFD_EBV"][starindices] 
+    else : 
+        log.warning("NO EXTINCTION VALUES IN FIBERMAP!!")
+    
+    
     # DIVIDE FLAT AND SUBTRACT SKY , TRIM DATA
     ############################################
     for cam in frames :
@@ -181,21 +189,24 @@ def main(args) :
             log.warning("Missing flat for %s"%cam)
             frames.pop(cam)
             continue
+        
 
-        frames[cam].flux = frames[cam].flux[starindices]
-        frames[cam].ivar = frames[cam].ivar[starindices]
-
-
-        frames[cam].ivar *= (frames[cam].mask[starindices] == 0)
-        frames[cam].ivar *= (skies[cam].ivar[starindices] != 0)
-        frames[cam].ivar *= (skies[cam].mask[starindices] == 0)
-        frames[cam].ivar *= (flats[cam].ivar[starindices] != 0)
-        frames[cam].ivar *= (flats[cam].mask[starindices] == 0)
-        frames[cam].flux *= ( frames[cam].ivar > 0) # just for clean plots
-        for star in range(frames[cam].flux.shape[0]) :
-            ok=np.where((frames[cam].ivar[star]>0)&(flats[cam].fiberflat[star]!=0))[0]
-            if ok.size > 0 :
-                frames[cam].flux[star] = frames[cam].flux[star]/flats[cam].fiberflat[star] - skies[cam].flux[star]
+        flat=flats[cam]
+        for frame,sky in zip(frames[cam],skies[cam]) :
+            frame.flux = frame.flux[starindices]
+            frame.ivar = frame.ivar[starindices]
+            frame.ivar *= (frame.mask[starindices] == 0)
+            frame.ivar *= (sky.ivar[starindices] != 0)
+            frame.ivar *= (sky.mask[starindices] == 0)
+            frame.ivar *= (flat.ivar[starindices] != 0)
+            frame.ivar *= (flat.mask[starindices] == 0)
+            frame.flux *= ( frame.ivar > 0) # just for clean plots
+            for star in range(frame.flux.shape[0]) :
+                ok=np.where((frame.ivar[star]>0)&(flat.fiberflat[star]!=0))[0]
+                if ok.size > 0 :
+                    frame.flux[star] = frame.flux[star]/flat.fiberflat[star] - sky.flux[star]
+            frame.resolution_data = frame.resolution_data[starindices]
+        
     nstars = starindices.size
     starindices=None # we don't need this anymore
 
@@ -224,16 +235,42 @@ def main(args) :
         for m in range(stdflux.shape[0]) :
             model_mags[m,index]=filter_response.get_ab_magnitude(stdflux[m]*fluxunits,stdwave)
     log.info("done computing model mags")
+    
+    mean_extinction_delta_mags = None
+    mean_ebv = np.mean(ebv)
+    if mean_ebv > 0 :
+        log.info("Compute a mean delta_color from average E(B-V) = %3.2f based on canonial model star"%mean_ebv)
+        # compute a mean delta_color from mean_ebv based on canonial model star
+        #######################################################################
+        # will then use this color offset in the model pre-selection
+        # find canonical f-type model: Teff=6000, logg=4, Fe/H=-1.5
+        canonical_model = np.argmin((teff-6000.0)**2+(logg-4.0)**2+(feh+1.5)**2)
+        canonical_model_mags_without_extinction = model_mags[canonical_model]
+        canonical_model_mags_with_extinction    = np.zeros(canonical_model_mags_without_extinction.shape)
 
+        canonical_model_reddened_flux = stdflux[canonical_model]*dust_transmission(stdwave,mean_ebv)                
+        for index in range(len(model_filters)) :
+            if model_filters[index].startswith('WISE'):
+                log.warning('not computing stdstar {} mags'.format(model_filters[index]))
+                continue
+            filter_response=load_filter(model_filters[index])
+            canonical_model_mags_with_extinction[index]=filter_response.get_ab_magnitude(canonical_model_reddened_flux*fluxunits,stdwave)
+
+        mean_extinction_delta_mags = canonical_model_mags_with_extinction - canonical_model_mags_without_extinction
+         
+    
 
     # LOOP ON STARS TO FIND BEST MODEL
     ############################################
-    bestModelIndex=np.arange(nstars)
-    templateID=np.arange(nstars)
+    linear_coefficients=np.zeros((nstars,stdflux.shape[0]))
     chi2dof=np.zeros((nstars))
     redshift=np.zeros((nstars))
     normflux=[]
 
+
+    star_colors_array=np.zeros((nstars))
+    model_colors_array=np.zeros((nstars))
+    
     for star in range(nstars) :
 
         log.info("finding best model for observed star #%d"%star)
@@ -244,12 +281,15 @@ def main(args) :
         ivar = {}
         resolution_data = {}
         for camera in frames :
-            band=camera[0]
-            wave[band]=frames[camera].wave
-            flux[band]=frames[camera].flux[star]
-            ivar[band]=frames[camera].ivar[star]
-            resolution_data[band]=frames[camera].resolution_data[star]
 
+            for i,frame in enumerate(frames[camera]) :
+                identifier="%s-%d"%(camera,i)
+                wave[identifier]=frame.wave
+                flux[identifier]=frame.flux[star]
+                ivar[identifier]=frame.ivar[star]
+                resolution_data[identifier]=frame.resolution_data[star]
+        
+        
         # preselec models based on magnitudes
 
         # compute star color
@@ -259,6 +299,8 @@ def main(args) :
         filter1=imaging_filters[star][index1]
         filter2=imaging_filters[star][index2]
         star_color=imaging_mags[star][index1]-imaging_mags[star][index2]
+        star_colors_array[star]=star_color
+        
 
         # compute models color
         model_index1=-1
@@ -273,20 +315,57 @@ def main(args) :
             log.error("cannot compute '%s' model color from %s"%(color_name,filters))
         model_colors = model_mags[:,model_index1]-model_mags[:,model_index2]
 
+        # apply extinction here
+        # use the colors derived from the cannonical model with the mean ebv of the stars
+        # and simply apply a scaling factor based on the ebv of this star
+        # this is sufficiently precise for the broad model pre-selection we are doing here
+        # the exact reddening of the star to each pre-selected model is 
+        # apply afterwards
+        if mean_extinction_delta_mags is not None  and mean_ebv != 0 :
+            delta_color = ( mean_extinction_delta_mags[model_index1] - mean_extinction_delta_mags[model_index2] ) * ebv[star]/mean_ebv 
+            model_colors += delta_color
+            log.info("Apply a %s-%s color offset = %4.3f to the models for star with E(B-V)=%4.3f"%(model_filters[model_index1],model_filters[model_index2],delta_color,ebv[star]))
         # selection
-        selection = np.where(np.abs(model_colors-star_color)<args.delta_color)[0]
-
+        
+        selection = np.abs(model_colors-star_color)<args.delta_color
+        # smallest cube in parameter space including this selection (needed for interpolation)
+        new_selection = (teff>=np.min(teff[selection]))&(teff<=np.max(teff[selection]))
+        new_selection &= (logg>=np.min(logg[selection]))&(logg<=np.max(logg[selection]))
+        new_selection &= (feh>=np.min(feh[selection]))&(feh<=np.max(feh[selection]))
+        selection = np.where(new_selection)[0]
+        
+        
         log.info("star#%d fiber #%d, %s = %s-%s = %f, number of pre-selected models = %d/%d"%(star,starfibers[star],args.color,filter1,filter2,star_color,selection.size,stdflux.shape[0]))
+        
+        #index_in_selection,second_index_in_selection,frac,redshift[star],chi2dof[star]=match_templates(wave,flux,ivar,resolution_data,stdwave,stdflux[selection], teff[selection], logg[selection], feh[selection], ncpu=args.ncpu,z_max=args.z_max,z_res=args.z_res,template_error=args.template_error)
+        
+        # apply extinction to selected_models
+        dust_transmission_of_this_star = dust_transmission(stdwave,ebv[star])
+        selected_reddened_stdflux = stdflux[selection]*dust_transmission_of_this_star
+        
+        coefficients,redshift[star],chi2dof[star]=match_templates(wave,flux,ivar,resolution_data,stdwave,selected_reddened_stdflux, teff[selection], logg[selection], feh[selection], ncpu=args.ncpu,z_max=args.z_max,z_res=args.z_res,template_error=args.template_error)
+        
+        linear_coefficients[star,selection] = coefficients
+        
+        log.info('Star Fiber: {0}; TEFF: {1}; LOGG: {2}; FEH: {3}; Redshift: {4}; Chisq/dof: {5}'.format(starfibers[star],np.inner(teff,linear_coefficients[star]),np.inner(logg,linear_coefficients[star]),np.inner(feh,linear_coefficients[star]),redshift[star],chi2dof[star]))
+        
 
-        index_in_selection,redshift[star],chi2dof[star]=match_templates(wave,flux,ivar,resolution_data,stdwave,stdflux[selection], teff[selection], logg[selection], feh[selection], ncpu=args.ncpu,z_max=args.z_max,z_res=args.z_res)
-
-        bestModelIndex[star] = selection[index_in_selection]
-
-        log.info('Star Fiber: {0}; TemplateID: {1}; Redshift: {2}; Chisq/dof: {3}'.format(starfibers[star],bestModelIndex[star],redshift[star],chi2dof[star]))
         # Apply redshift to original spectrum at full resolution
-        tmp=np.interp(stdwave,stdwave/(1+redshift[star]),stdflux[bestModelIndex[star]])
+        model=np.zeros(stdwave.size)
+        for i,c in enumerate(linear_coefficients[star]) :
+            if c != 0 :
+                model += c*np.interp(stdwave,stdwave*(1+redshift[star]),stdflux[i])
+
+        # Apply dust extinction
+        model *= dust_transmission_of_this_star
+        
+        # Compute final model color
+        mag1=load_filter(model_filters[model_index1]).get_ab_magnitude(model*fluxunits,stdwave)
+        mag2=load_filter(model_filters[model_index2]).get_ab_magnitude(model*fluxunits,stdwave)
+        model_colors_array[star] = mag1-mag2        
+
         # Normalize the best model using reported magnitude
-        normalizedflux=normalize_templates(stdwave,tmp,imaging_mags[star],imaging_filters[star])
+        normalizedflux=normalize_templates(stdwave,model,imaging_mags[star],imaging_filters[star])
         normflux.append(normalizedflux)
 
 
@@ -294,9 +373,14 @@ def main(args) :
     # Now write the normalized flux for all best models to a file
     normflux=np.array(normflux)
     data={}
-    data['BESTMODEL']=bestModelIndex
-    data['TEMPLATEID']=bestModelIndex # IS THAT IT?
+    data['LOGG']=linear_coefficients.dot(logg)
+    data['TEFF']= linear_coefficients.dot(teff)
+    data['FEH']= linear_coefficients.dot(feh)
     data['CHI2DOF']=chi2dof
     data['REDSHIFT']=redshift
+    data['COEFF']=linear_coefficients
+    data['DATA_%s'%args.color]=star_colors_array
+    data['MODEL_%s'%args.color]=model_colors_array
     norm_model_file=args.outfile
     io.write_stdstar_models(args.outfile,normflux,stdwave,starfibers,data)
+